+++ deviceappmanagement/mobile_apps_graph_managed_mobile_lob_app_request_builder.go
@@ -29,9 +29,9 @@ type MobileAppsGraphManagedMobileLobAppRequestBuilderGetQueryParameters struct {
     // Select properties to be returned
     Select []string `uriparametername:"%24select"`
     // Skip the first n items
-    Skip *int32 `uriparametername:"%24skip"`
+    Skip *int64 `uriparametername:"%24skip"`
     // Show only the first n items
-    Top *int32 `uriparametername:"%24top"`
+    Top *int64 `uriparametername:"%24top"`
 }
 // MobileAppsGraphManagedMobileLobAppRequestBuilderGetRequestConfiguration configuration for the request such as headers, query parameters, and middleware options.
 type MobileAppsGraphManagedMobileLobAppRequestBuilderGetRequestConfiguration struct {
