+++ drives/item_items_item_search_with_q_request_builder.go
@@ -28,9 +28,9 @@ type ItemItemsItemSearchWithQRequestBuilderGetQueryParameters struct {
     // Select properties to be returned
     Select []string `uriparametername:"%24select"`
     // Skip the first n items
-    Skip *int32 `uriparametername:"%24skip"`
+    Skip *int64 `uriparametername:"%24skip"`
     // Show only the first n items
-    Top *int32 `uriparametername:"%24top"`
+    Top *int64 `uriparametername:"%24top"`
 }
 // ItemItemsItemSearchWithQRequestBuilderGetRequestConfiguration configuration for the request such as headers, query parameters, and middleware options.
 type ItemItemsItemSearchWithQRequestBuilderGetRequestConfiguration struct {
@@ -61,9 +61,6 @@ func NewItemItemsItemSearchWithQRequestBuilder(rawUrl string, requestAdapter i2a
 // Deprecated: This method is obsolete. Use GetAsSearchWithQGetResponse instead.
 // returns a ItemItemsItemSearchWithQResponseable when successful
 // returns a ODataError error when the service returns a 4XX or 5XX status code
-// [Find more info here]
-// 
-// [Find more info here]: https://learn.microsoft.com/graph/api/driveitem-search?view=graph-rest-1.0
 func (m *ItemItemsItemSearchWithQRequestBuilder) Get(ctx context.Context, requestConfiguration *ItemItemsItemSearchWithQRequestBuilderGetRequestConfiguration)(ItemItemsItemSearchWithQResponseable, error) {
     requestInfo, err := m.ToGetRequestInformation(ctx, requestConfiguration);
     if err != nil {
@@ -84,9 +81,6 @@ func (m *ItemItemsItemSearchWithQRequestBuilder) Get(ctx context.Context, reques
 // GetAsSearchWithQGetResponse search the hierarchy of items for items matching a query.You can search within a folder hierarchy, a whole drive, or files shared with the current user.
 // returns a ItemItemsItemSearchWithQGetResponseable when successful
 // returns a ODataError error when the service returns a 4XX or 5XX status code
-// [Find more info here]
-// 
-// [Find more info here]: https://learn.microsoft.com/graph/api/driveitem-search?view=graph-rest-1.0
 func (m *ItemItemsItemSearchWithQRequestBuilder) GetAsSearchWithQGetResponse(ctx context.Context, requestConfiguration *ItemItemsItemSearchWithQRequestBuilderGetRequestConfiguration)(ItemItemsItemSearchWithQGetResponseable, error) {
     requestInfo, err := m.ToGetRequestInformation(ctx, requestConfiguration);
     if err != nil {
